[MONOREPO]: Moving catalogd's metrics test to operator-controller tests
